Refactor MockHttpRequest.
