+++ storage/client/storage_client_testing.bzl
@@ -1,0 +1,9 @@
+# DO NOT EDIT -- GENERATED BY CMake -- Change the CMakeLists.txt file if needed
+storage_client_testing_HDRS = [
+    "testing/mock_http_request.h",
+]
+
+storage_client_testing_SRCS = [
+    "testing/mock_http_request.cc",
+]
+
